% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/address_lokup.r
\name{address_lookup}
\alias{address_lookup}
\title{Lookup the address of one or multiple OSM objects like node, way or relation.}
\usage{
address_lookup(osm_ids, address_details = TRUE,
  email = getOption("OSM_API_EMAIL", "nominatimrpackage@example.com"),
  accept_language = getOption("LANG", "en-US,en;q=0.8"))
}
\arguments{
\item{osm_ids}{A vector of up to 50 specific osm node, way or relations ids to
return the addresses for. Format for each entry is \code{[N|W|R]<value>}.}

\item{address_details}{Include a breakdown of the address into elements (TRUE == include)}

\item{email}{If you are making large numbers of request please include a valid email address
or alternatively include your email address as part of the User-Agent string.
This information will be kept confidential and only used to contact you in the
event of a problem, see Usage Policy for more details. You can pass the value
in directly or set the \code{OSM_API_EMAIL} option and the function will
it for all requests.}

\item{accept_language}{Preferred language order for showing search results
Either uses standard rfc2616 accept-language string or a simple comma separated l
ist of language codes. The \code{LANG} option will be used, if set.#'}
}
\description{
Vectorized over \code{osm_ids}
}
\examples{
\dontrun {
places <- c("R146656", "W104393803", "N240109189")
places_found <- address_lookup(places)
}
}

