% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_lookup.r
\name{btlr_to_bb}
\alias{btlr_to_bb}
\title{Convert metadata generated by \link{bb_lookup} to bounding box}
\usage{
btlr_to_bb(btlr)
}
\arguments{
\item{btlr}{metadata provided by \link{bb_lookup}}
}
\value{
\link[sp]{bbox}
}
\description{
Convert metadata generated by \link{bb_lookup} to bounding box
}
\examples{
meta <- bb_lookup("United States")
btlr <- meta[1,c("bottom", "top", "left", "right")]
btlr_to_bb(btlr)
}
