% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bb_lookup.r
\name{bb_lookup}
\alias{bb_lookup}
\title{Bounding box (and other metadata) lookup}
\usage{
bb_lookup(query, viewbox = NULL)
}
\arguments{
\item{query}{search terms to pass to OSM Nominatim}

\item{viebox}{(optional) atomic character vector with comma-separated values
              (e.g. \code{"-4.37,54.88,2.04,52.96"}) or a 4-element vector
              specifying the viewbox (e.g. \code{c(-4.37,54.88,2.04,52.96)}).}
}
\value{
\code{data.frame} of places with metatata, including bounding box information
}
\description{
Perform a general \href{https://nominatim.openstreetmap.org/}{Nominatim} query
and retrieve \code{place} metadata, including bounding box information
}
\examples{
bb_lookup("West Yorkshire", c(-4.37, 54.88, 2.04, 52.96))
bb_lookup("United States")
}

