% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nominatim-package.r
\docType{package}
\name{nominatim}
\alias{nominatim}
\alias{nominatim-package}
\title{An interface between R and the OpenStreeMap Nominatim API}
\description{
Most queries are run against the MapQuest Nominatim API as per the OpenStreetMap
recommendation.
}
\details{
Data (c) OpenStreetMap contributors, ODbL 1.0. \url{http://www.openstreetmap.org/copyright}\cr
Nominatim Usage Policy: \url{http://wiki.openstreetmap.org/wiki/Nominatim_usage_policy}\cr
MapQuest Nominatim Terms of Use: \url{http://info.mapquest.com/terms-of-use/}\cr

\url{http://wiki.openstreetmap.org/wiki/Nominatim}
}
\note{
A slight delay is introduced between calls to both OpenStreetMap Nominatim &
      MapQuest Nominatim API to reduce load on their servers.
}
\author{
Bob Rudis (@hrbrmstr)
}
