% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nominatim-package.r
\docType{package}
\name{nominatim}
\alias{nominatim}
\alias{nominatim-package}
\title{A package to is an R package to interface to the OpenStreeMap Nominatum API
\url{http://wiki.openstreetmap.org/wiki/Nominatim}}
\description{
A package to is an R package to interface to the OpenStreeMap Nominatum API
\url{http://wiki.openstreetmap.org/wiki/Nominatim}
}
\author{
Bob Rudis (@hrbrmstr)
}

